+++ packages/ports/Cargo.toml
@@ -10,11 +10,12 @@ ethers-core = { workspace = true, optional = true }
 futures = { workspace = true, optional = true }
 impl-tools = { workspace = true, optional = true }
 mockall = { workspace = true, optional = true }
+rand = { workspace = true, optional = true }
+serde = { workspace = true, features = ["derive"] }
 thiserror = { workspace = true, optional = true }
-types = { workspace = true }
 
 [features]
-test-helpers = ["dep:mockall", "types/test-helpers"]
+test-helpers = ["dep:mockall", "dep:rand"]
 eth = ["dep:ethers-core", "dep:futures", "dep:thiserror", "dep:async-trait"]
 fuel = ["dep:thiserror", "dep:async-trait"]
 storage = ["dep:impl-tools", "dep:thiserror", "dep:async-trait"]
