+++ packages/ports/src/ports/eth_rpc.rs
@@ -1,0 +1,36 @@
+use std::pin::Pin;
+
+use crate::types::{EthHeight, FuelBlock, FuelBlockCommittedOnEth, InvalidEthHeight, Stream, U256};
+
+#[derive(Debug, thiserror::Error)]
+pub enum Error {
+    #[error("network error: {0}")]
+    Network(String),
+    #[error("{0}")]
+    Other(String),
+}
+
+pub type Result<T> = std::result::Result<T, Error>;
+
+impl From<InvalidEthHeight> for Error {
+    fn from(err: InvalidEthHeight) -> Self {
+        Self::Other(err.to_string())
+    }
+}
+
+#[cfg_attr(feature = "test-helpers", mockall::automock)]
+#[async_trait::async_trait]
+pub trait EthereumAdapter: Send + Sync {
+    async fn submit(&self, block: FuelBlock) -> Result<()>;
+    async fn get_block_number(&self) -> Result<EthHeight>;
+    async fn balance(&self) -> Result<U256>;
+    fn event_streamer(&self, eth_block_height: u64) -> Box<dyn EventStreamer + Send + Sync>;
+}
+
+#[cfg_attr(feature = "test-helpers", mockall::automock)]
+#[async_trait::async_trait]
+pub trait EventStreamer {
+    async fn establish_stream<'a>(
+        &'a self,
+    ) -> Result<Pin<Box<dyn Stream<Item = Result<FuelBlockCommittedOnEth>> + 'a + Send>>>;
+}
