+++ packages/ports/src/ports/fuel_rpc.rs
@@ -1,3 +1,5 @@
+use crate::types::FuelBlock;
+
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
     #[error("{0}")]
@@ -6,9 +8,9 @@ pub enum Error {
 
 pub type Result<T> = std::result::Result<T, Error>;
 
-#[async_trait::async_trait]
 #[cfg_attr(feature = "test-helpers", mockall::automock)]
+#[async_trait::async_trait]
 pub trait FuelAdapter: Send + Sync {
-    async fn block_at_height(&self, height: u32) -> Result<Option<crate::FuelBlock>>;
-    async fn latest_block(&self) -> Result<crate::FuelBlock>;
+    async fn block_at_height(&self, height: u32) -> Result<Option<FuelBlock>>;
+    async fn latest_block(&self) -> Result<FuelBlock>;
 }
