+++ packages/ports/src/ports/storage.rs
@@ -1,5 +1,7 @@
 use std::sync::Arc;
 
+use crate::types::BlockSubmission;
+
 #[derive(Debug, thiserror::Error)]
 #[error("{msg}")]
 pub struct Error {
@@ -19,10 +21,7 @@ pub type Result<T> = std::result::Result<T, Error>;
 #[impl_tools::autoimpl(for<T: trait> &T, &mut T, Arc<T>, Box<T>)]
 #[cfg_attr(feature = "test-helpers", mockall::automock)]
 pub trait Storage: Send + Sync {
-    async fn insert(&self, submission: crate::BlockSubmission) -> Result<()>;
-    async fn submission_w_latest_block(&self) -> Result<Option<crate::BlockSubmission>>;
-    async fn set_submission_completed(
-        &self,
-        fuel_block_hash: [u8; 32],
-    ) -> Result<crate::BlockSubmission>;
+    async fn insert(&self, submission: BlockSubmission) -> Result<()>;
+    async fn submission_w_latest_block(&self) -> Result<Option<BlockSubmission>>;
+    async fn set_submission_completed(&self, fuel_block_hash: [u8; 32]) -> Result<BlockSubmission>;
 }
