+++ packages/ports/src/types.rs
@@ -1,0 +1,16 @@
+#[cfg(feature = "eth")]
+pub use ethers_core::types::{H160, U256};
+#[cfg(feature = "eth")]
+pub use futures::Stream;
+
+mod block_submission;
+mod eth_height;
+mod fuel_block;
+#[cfg(feature = "eth")]
+mod fuel_block_committed_on_eth;
+
+pub use block_submission::*;
+pub use eth_height::*;
+pub use fuel_block::*;
+#[cfg(feature = "eth")]
+pub use fuel_block_committed_on_eth::*;
