+++ packages/ports/src/types/fuel_block.rs
