+++ packages/storage/Cargo.toml
@@ -23,7 +23,7 @@ tokio = { workspace = true, optional = true }
 [dev-dependencies]
 ports = { workspace = true, features = ["storage"] }
 rand = { workspace = true }
-storage = { path = ".", features = ["test-helpers"] }
+storage = { workspace = true, features = ["test-helpers"] }
 tokio = { workspace = true }
 
 [features]
