--- packages/types/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "types"
-version = "0.1.0"
-edition = "2021"
-publish = false
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-rand = { workspace = true, optional = true }
-serde = { workspace = true, features = ["derive"] }
-
-[features]
-test-helpers = ["dep:rand"]
