--- packages/types/src/lib.rs
@@ -1,8 +1,0 @@
-#![deny(unused_crate_dependencies)]
-mod block_submission;
-mod eth_height;
-mod fuel_block;
-
-pub use block_submission::*;
-pub use eth_height::*;
-pub use fuel_block::*;
